+++ account_payment_term_manual_sale/models/__init__.py
@@ -1,2 +1,1 @@
-# from . import sale_payment_term_manual
-from . import sale_order
+from . import sale_payment_term_manual
