+++ account_payment_term_manual_sale/models/sale_payment_term_manual.py
@@ -1,17 +1,28 @@
 # Copyright 2024 KMEE
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
-from odoo import fields, models
+from odoo import api, models
 
 
-class SalePaymentTermManual(models.Model):
+class SaleOrder(models.Model):
 
-    _name = "sale.payment.term.manual"
-    _inherit = "account.payment.term.manual.mixin"
+    _name = "sale.order"
+    _inherit = [
+        "sale.order",
+        "account.payment.term.manual.mixin",
+    ]
 
-    sale_order_id = fields.Many2one("sale.order")
-    currency_id = fields.Many2one(
-        "res.currency",
-        related="sale_order_id.currency_id",
-        readonly=True,
-    )
+    @api.onchange("payment_term_id")
+    def _onchange_payment_term_id(self):
+        if not self.payment_term_id:
+            return
+        # Replace manual lines if:
+        # 1. invoice term has changed directly
+        # 2. invoice term has changed indirectly but no current manual term is set
+        if (
+            self.env.context.get("payment_term_id_view_onchange")
+            or not self.manual_payment_term_id
+        ):
+            new_term_id = self.payment_term_id
+            self._update_manual_payment_term_id(self.payment_term_id)
+            self.payment_term_id = new_term_id
